requirements.txt
emoji==1.7.0
Flask==2.1.1
Jinja2==3.1.1
matplotlib==3.5.1
numpy==1.21.6
Pillow==9.1.0
python-dateutil==2.8.2
Flask-Session
imgkit
Flask-WeasyPrint